Add profileLayout and move personal data page to nested route
